:bug: Only enable touch events when necessary
Previous versions of this library subscribed to both touch and mouse
events at the same time, which caused the switch to trigger 2 events at
once when using mobile devices with support to both click and touch
events.

This turned out to be a problem in switches that were wide enough that
would capture more than one event at once, causing the component to turn
the switch on and off at the same time.